mesh_device dedicated RT module
