+++ tt_metal/distributed/mesh_device.cpp
@@ -21,7 +21,6 @@
 #include <algorithm>
 #include <array>
 #include <atomic>
-#include <cmath>
 #include <cstddef>
 #include <iterator>
 #include <memory>
@@ -32,10 +31,7 @@
 #include "impl/allocator/allocator.hpp"
 #include <tt_stl/assert.hpp>
 #include "buffer.hpp"
-#include <tt-metalium/mesh_buffer.hpp>
 #include "device/device_impl.hpp"
-#include "dispatch/command_queue_common.hpp"
-#include "dispatch/dispatch_mem_map.hpp"
 #include "dispatch/dispatch_settings.hpp"
 #include "host_api.hpp"
 #include "mesh_config.hpp"
@@ -47,23 +43,16 @@
 #include <tt_stl/strong_type.hpp>
 #include "impl/threading/thread_pool.hpp"
 #include "device/device_manager.hpp"
-#include <enchantum/enchantum.hpp>
 #include <experimental/fabric/control_plane.hpp>
 #include <experimental/fabric/fabric_types.hpp>
 #include "distributed/fd_mesh_command_queue.hpp"
+#include "distributed/realtime_profiler_manager.hpp"
 #include "distributed/sd_mesh_command_queue.hpp"
 #include "tracy/Tracy.hpp"
-#include <common/TracySystem.hpp>
-#include <tracy/TracyTTDevice.hpp>
-#include <common/TracyTTDeviceData.hpp>
 #include "tools/profiler/tt_metal_tracy.hpp"
 #include <env_lib.hpp>
-#include "llrt/hal.hpp"
 
 #include "allocator/l1_banking_allocator.hpp"
-#include "tt_metal/impl/dispatch/data_collection.hpp"
-#include "tt_metal/impl/dispatch/kernels/realtime_profiler_ring_buffer.hpp"
-#include "tt_metal/impl/dispatch/realtime_profiler_tracy_handler.hpp"
 #include "debug/inspector/inspector.hpp"
 #include "sub_device/sub_device_manager.hpp"
 #include "sub_device/sub_device_manager_tracker.hpp"
@@ -160,200 +149,6 @@ MeshCoordinate compute_system_mesh_offset(const MeshDeviceView& view) {
         }
     }
     TT_THROW("Failed to find offset for mesh device view");
-}
-
-// Real-time profiler runtime constants. Sizes that govern the on-device L1 layout
-// live in tt_metal/impl/dispatch/kernels/realtime_profiler_ring_buffer.hpp
-// (RealtimeProfilerCoreL1, RT_PROFILER_RING_CAPACITY, RT_PROFILER_ENTRY_SIZE) and
-// are reused here directly so the host and device agree on a single source of truth.
-struct RealtimeProfilerRuntimeSizes {
-    static constexpr uint32_t fifo_size = 4096;                    // 4KB pinned-host FIFO for D2H socket
-    static constexpr uint32_t page_size = RT_PROFILER_ENTRY_SIZE;  // host page size == ring entry size
-    static constexpr uint32_t core_l1_size = sizeof(RealtimeProfilerCoreL1);
-};
-
-// Compute the L1 carve-out addresses on the reserved RT-profiler tensix core for a given
-// RealtimeProfilerCoreL1 base. Anchors at dispatch_mem_map's UNRESERVED so we sit cleanly
-// after every dispatch carve-out, avoiding the user-space allocator entirely.
-inline RealtimeProfilerCoreL1Addrs compute_rt_profiler_core_l1_addrs(uint32_t base) {
-    return {
-        .base = base,
-        .ring_buffer = base + static_cast<uint32_t>(offsetof(RealtimeProfilerCoreL1, ring)),
-        .socket_config = base + static_cast<uint32_t>(offsetof(RealtimeProfilerCoreL1, socket_config)),
-    };
-}
-
-// Result of evaluating whether the real-time profiler can be brought up on a given device.
-struct RealtimeProfilerEligibility {
-    bool enabled = false;
-    CoreCoord core;  // Only meaningful when enabled == true; tensix core reserved for the RT profiler kernel.
-};
-
-// Consolidated eligibility check for the real-time profiler on a single device.
-//
-// Logs (at info/warning) the reason RT profiler is being disabled so the user gets a
-// clear explanation, and returns {enabled=false} in that case. When enabled, the
-// returned `core` is the dispatch-pool-reserved tensix to use both for the RT
-// profiler BRISC kernel and as the sender core of the D2H socket.
-//
-// Checks performed (in order):
-//   1. Device is MMIO-capable. D2H sockets need a PCIe-connected sender core;
-//      remote devices (e.g. device 1 on N300) cannot map pinned host memory.
-//   2. The D2H socket's memory-allocation path is supported on this host. On
-//      architectures with 64-bit PCIe addressing (Blackhole) the D2H socket
-//      only takes the pinned-memory path — there is no hugepage fallback — so
-//      IOMMU must be enabled on the host for the pinning to succeed. If not,
-//      we disable cleanly rather than fault inside UMD's sysmem mapping.
-//   3. Fabric tensix datamover (MUX / UDM) is disabled. When it's enabled, the
-//      fabric mux drains the dispatch-pool at fabric-init time — on small-pool
-//      MMIO chips (e.g. T3K's FABRIC_2D + UDM path) a single extra tensix
-//      reserved for RT profiler tips the pool into exhaustion and makes
-//      fabric_mux_core() TT_THROW "No more available dispatch cores". RT
-//      profiler yields its slot in that case.
-//   4. A tensix core was reserved for the RT profiler at dispatch_core_manager
-//      construction time. The reservation only happens for MMIO chips with
-//      WORKER dispatch — for ETH dispatch the dispatch pool holds ethernet
-//      cores, which cannot run the RT profiler's BRISC worker kernel. Check 3
-//      has already handled the fabric-mux branch of the same reservation.
-//   5. Defensive: the reserved coordinate lives inside the logical TENSIX grid
-//      (harvesting overlays could in principle produce a stale entry).
-//   6. Kernels are not nullified (DEBUG_NULL_KERNELS / TT_METAL_NULL_KERNELS).
-//      When active, the JIT replaces every non-dispatch kernel body — including
-//      the RT profiler BRISC/NCRISC — with a stub that returns immediately
-//      after wait_for_go_message(). The RT profiler state machine would never
-//      run, so every host sync would time out (~40s wasted per MeshDevice).
-//      Both host and device profiling are meaningless in this mode anyway
-//      because the user kernels being timed are themselves stubs.
-//   7. The L1 bank on the reserved profiler core has room for the
-//      ~256 KB BRISC↔NCRISC ring buffer. Fixtures that intentionally shrink
-//      `worker_l1_size` to stress the kernel-config region (e.g. the dispatch
-//      KernelSizeTestBigBuffer suite, which sets it to 64 KB) leave too little
-//      user-allocatable L1 for the ring; we disable cleanly instead of
-//      OOM-faulting in MeshDevice setup.
-RealtimeProfilerEligibility evaluate_realtime_profiler_eligibility(IDevice* device) {
-    auto device_id = device->id();
-    auto& metal = MetalContext::instance();
-    const auto& hal = metal.hal();
-    const auto& cluster = metal.get_cluster();
-    auto& dispatch_core_manager = metal.get_dispatch_core_manager();
-
-    // 1. MMIO capability.
-    if (!device->is_mmio_capable()) {
-        log_info(
-            tt::LogMetal,
-            "Real-time profiler disabled on device {}: device is not MMIO-capable (remote device). "
-            "D2H sockets require the sender core to sit on a PCIe-connected chip.",
-            device_id);
-        return {};
-    }
-
-    // 2. D2H socket memory-allocation precondition (IOMMU vs. hugepage fallback).
-    if (hal.get_supports_64_bit_pcie_addressing() && !cluster.is_iommu_enabled()) {
-        log_info(
-            tt::LogMetal,
-            "Real-time profiler disabled on device {}: this architecture uses 64-bit PCIe "
-            "addressing for the D2H socket, which requires IOMMU to be enabled on the host. "
-            "IOMMU is currently disabled and no hugepage fallback is available. Enable IOMMU "
-            "(or run on a system that has it) to re-enable RT profiler.",
-            device_id);
-        return {};
-    }
-
-    // 3. Fabric tensix datamover (MUX / UDM) competes for the same dispatch-core pool.
-    //    Mirrors the skip path in dispatch_core_manager::reset_dispatch_core_manager — keep
-    //    both sites in sync.
-    const auto fabric_tensix_config = metal.get_fabric_tensix_config();
-    if (fabric_tensix_config != tt_fabric::FabricTensixConfig::DISABLED) {
-        log_info(
-            tt::LogMetal,
-            "Real-time profiler disabled on device {}: fabric tensix datamover is enabled "
-            "(FabricTensixConfig={}, FabricUDMMode={}), and fabric_mux_core() will drain the "
-            "remaining dispatch-pool cores at fabric-init time. Reserving a tensix for the RT "
-            "profiler on top of that tips the pool into exhaustion on small-pool chips. "
-            "Disable the fabric tensix datamover to re-enable RT profiler.",
-            device_id,
-            enchantum::to_string(fabric_tensix_config),
-            enchantum::to_string(metal.get_fabric_udm_mode()));
-        return {};
-    }
-
-    // 4. Reserved tensix core for the RT profiler kernel / D2H sender. When absent here
-    //    (with the fabric-mux path already ruled out above) the cause is ETH dispatch:
-    //    the dispatch pool holds ethernet cores, which can't run the RT profiler BRISC.
-    std::optional<tt_cxy_pair> reserved = dispatch_core_manager.get_reserved_realtime_profiler_core(device_id);
-    if (!reserved.has_value()) {
-        log_info(
-            tt::LogMetal,
-            "Real-time profiler disabled on device {}: no tensix core could be reserved for the "
-            "RT profiler. Dispatch is configured for ETH cores, which cannot run the RT profiler "
-            "BRISC kernel. Switch to DispatchCoreConfig(DispatchCoreType::WORKER) to re-enable RT "
-            "profiler.",
-            device_id);
-        return {};
-    }
-
-    CoreCoord core(reserved->x, reserved->y);
-
-    // 5. Sanity-check the reservation against the TENSIX grid.
-    const auto& soc = cluster.get_soc_desc(device_id);
-    CoreCoord tensix_grid = soc.get_grid_size(CoreType::TENSIX);
-    if (core.x >= tensix_grid.x || core.y >= tensix_grid.y) {
-        log_warning(
-            tt::LogMetal,
-            "Real-time profiler disabled on device {}: reserved core ({}, {}) is outside the "
-            "TENSIX logical grid ({}, {}).",
-            device_id,
-            core.x,
-            core.y,
-            tensix_grid.x,
-            tensix_grid.y);
-        return {};
-    }
-
-    // 6. Null-kernels mode: the RT profiler's BRISC/NCRISC bodies would be
-    //    stubbed out by -DDEBUG_NULL_KERNELS, so the on-device state machine
-    //    never runs and every host sync would time out. Skip cleanly.
-    if (metal.rtoptions().get_kernels_nullified()) {
-        log_info(
-            tt::LogMetal,
-            "Real-time profiler disabled on device {}: null-kernels mode is active "
-            "(TT_METAL_NULL_KERNELS / set_kernels_nullified). The RT profiler kernel "
-            "would be replaced with a stub and could not respond to host syncs, and "
-            "there are no real user kernels to profile in this mode.",
-            device_id);
-        return {};
-    }
-
-    // 7. L1 bank on the reserved profiler core must have room for the full RT-profiler
-    //    L1 layout (ring buffer + D2H socket sender config). The layout is carved out
-    //    of dispatch L1 starting at the post-UNRESERVED boundary; the L1 bank size is a
-    //    conservative proxy for "user-addressable L1 past the dispatch carve-outs", so a
-    //    bank that doesn't fit the layout means the carve-out will run past L1 too.
-    //    Fixtures that intentionally shrink worker_l1_size to stress the kernel-config
-    //    region (e.g. the dispatch KernelSizeTestBigBuffer suite, which sets it to
-    //    64 KB) leave too little user-allocatable L1 for the layout; skip cleanly.
-    const uint32_t l1_alignment = hal.get_alignment(HalMemType::L1);
-    const uint32_t core_l1_size_aligned = tt::align(RealtimeProfilerRuntimeSizes::core_l1_size, l1_alignment);
-    const DeviceAddr l1_bank_size = device->allocator()->get_bank_size(BufferType::L1);
-    if (l1_bank_size < core_l1_size_aligned) {
-        log_warning(
-            tt::LogMetal,
-            "Real-time profiler disabled on device {}: not enough user-allocatable L1 on the "
-            "reserved profiler core ({}, {}) for the RT-profiler L1 layout "
-            "(need {} B, L1 bank size is {} B). This typically means worker_l1_size was set "
-            "well below the default (e.g. a stress test pinning it at 64 KB to maximize the "
-            "kernel-config region). Increase worker_l1_size by at least {} B (or leave it at "
-            "the default) to re-enable RT profiler.",
-            device_id,
-            core.x,
-            core.y,
-            core_l1_size_aligned,
-            l1_bank_size,
-            core_l1_size_aligned - l1_bank_size);
-        return {};
-    }
-
-    return {.enabled = true, .core = core};
 }
 
 }  // namespace
@@ -1066,10 +861,8 @@ bool MeshDeviceImpl::close_impl(MeshDevice* pimpl_wrapper) {
 
     log_trace(tt::LogMetal, "Closing mesh device {}", this->id());
 
-    // Shut down the CQ FIRST so dispatch_s processes TERMINATE and signals
-    // the profiler core with the final buffer.  The profiler infrastructure
-    // (push kernel, receiver thread, callback) must still be alive at this
-    // point to deliver the last record to the host.
+    // Shut down the CQ first so dispatch_s sends TERMINATE to the profiler core with the
+    // final buffer; the push kernel, receiver thread, and callbacks must still be alive.
     if (is_initialized()) {
         if (MetalContext::instance(this->get_context_id()).get_cluster().get_target_device_type() !=
             tt::TargetDevice::Mock) {
@@ -1120,48 +913,12 @@ bool MeshDeviceImpl::close_impl(MeshDevice* pimpl_wrapper) {
         mesh_command_queues_.clear();
     }
 
-    // CQ shutdown is complete — dispatch_s has signaled the last profiler
-    // buffer and sent TERMINATE to the profiler core.  The profiler core
-    // set ring_buffer->terminate, so the push kernel will drain the ring
-    // buffer and exit.  Write the terminate flag again as a safety net,
-    // then give the push kernel time to deliver the final PCIe page.
-    for (auto& dev_state : realtime_profiler_devices_) {
-        if (dev_state.core_l1.ring_buffer != 0 && dev_state.device) {
-            const uint32_t terminate_addr = dev_state.core_l1.ring_buffer + offsetof(RtProfilerRingBuffer, terminate);
-            std::vector<uint32_t> terminate_flag = {1};
-            try {
-                tt::tt_metal::detail::WriteToDeviceL1(
-                    dev_state.device,
-                    dev_state.realtime_profiler_core,
-                    terminate_addr,
-                    terminate_flag,
-                    CoreType::WORKER);
-            } catch (const std::exception& e) {
-                log_warning(
-                    tt::LogMetal,
-                    "[Real-time profiler] Failed to write terminate flag for device {}: {}",
-                    dev_state.chip_id,
-                    e.what());
-            }
-        }
+    // Tear down RT profiler after the CQ has shut down (so dispatch_s has already issued
+    // the final TERMINATE) but before the rest of the device teardown.
+    if (realtime_profiler_) {
+        realtime_profiler_->shutdown();
+        realtime_profiler_.reset();
     }
-    if (!realtime_profiler_devices_.empty()) {
-        std::this_thread::sleep_for(std::chrono::milliseconds(100));
-    }
-
-    if (realtime_profiler_thread_.joinable()) {
-        realtime_profiler_stop_.store(true);
-        realtime_profiler_thread_.join();
-    }
-
-    realtime_profiler_tracy_handler_.reset();
-    // Clear activation state before destroying the per-device records so that a
-    // concurrent tt::IsProgramRealtimeProfilerActive() query never sees a stale
-    // "active" entry for a chip that is in the middle of shutdown.
-    for (const auto& dev_state : realtime_profiler_devices_) {
-        tt::NotifyProgramRealtimeProfilerDeactivated(dev_state.chip_id);
-    }
-    realtime_profiler_devices_.clear();
 
     if (is_initialized()) {
         sub_device_manager_tracker_.reset();
@@ -1623,963 +1380,21 @@ bool MeshDeviceImpl::initialize_impl(
     return true;
 }
 
-// Sync marker ID - must match device-side REALTIME_PROFILER_SYNC_MARKER_ID
-constexpr uint32_t REALTIME_PROFILER_SYNC_MARKER_ID = 0xFFFFFFFF;
-
-namespace {
-// Host clock wrapper for the RT profiler sync handshake.
-//
-// When TRACY_ENABLE is defined, TracyGetCpuTime() returns a real monotonic
-// host timestamp (hardware TSC or CLOCK_MONOTONIC_RAW ns) and
-// TracyGetTimerMul() converts ticks -> nanoseconds. When Tracy is
-// disabled, the Tracy header stubs both to literal 0, which breaks the
-// handshake: the host ends up writing sync_host_timestamp = 0 to L1, and
-// the device kernel's `if (host_time > 0)` guard silently drops the sync
-// request. The sync then times out and retries for minutes in CI.
-//
-// Fall back to std::chrono::steady_clock (nanoseconds) when Tracy is off,
-// so the handshake works regardless of build config. The Tracy-enabled
-// path is a pure pass-through, so behavior on Tracy builds is unchanged.
-inline int64_t rt_profiler_host_ticks() {
-#ifdef TRACY_ENABLE
-    return TracyGetCpuTime();
-#else
-    return std::chrono::duration_cast<std::chrono::nanoseconds>(std::chrono::steady_clock::now().time_since_epoch())
-        .count();
-#endif
-}
-
-inline double rt_profiler_ns_per_tick() {
-#ifdef TRACY_ENABLE
-    return TracyGetTimerMul();
-#else
-    // steady_clock already returns nanoseconds, so 1 tick == 1 ns.
-    return 1.0;
-#endif
-}
-}  // namespace
-
-void MeshDeviceImpl::run_realtime_profiler_sync(RealtimeProfilerDeviceState& dev_state, uint32_t num_samples) {
-    auto& cluster = MetalContext::instance().get_cluster();
-    int64_t host_start_time = rt_profiler_host_ticks();
-
-    struct SyncSample {
-        int64_t host_time;     // Full 64-bit host TSC ticks relative to host_start_time
-        uint64_t device_time;  // Device wall clock cycles
-    };
-    std::vector<SyncSample> samples;
-
-    // Discard any pre-existing pages from the socket before entering sync mode
-    // WITHOUT touching the data region.  The data region is mapped via PCIe and
-    // may contain undefined values on the very first sync of a freshly-
-    // constructed MeshDevice (the kernel may not have written valid data yet,
-    // or the shmem region may have been recycled by Linux from a prior
-    // process).  Reading it as a "stale page" and interpreting the contents as
-    // a sync response can corrupt the marker check downstream.
-    // discard_pending_pages() rebases bytes_acked -> bytes_sent and notifies
-    // the device, achieving the same flow-control effect without relying on
-    // the data region's contents.
-    constexpr uint32_t kSyncPageWords = 64 / sizeof(uint32_t);
-    uint32_t stale_pages = dev_state.socket->discard_pending_pages();
-    if (stale_pages > 0) {
-        log_info(
-            tt::LogMetal,
-            "[Real-time profiler] Device {} discarded {} stale pages before sync",
-            dev_state.chip_id,
-            stale_pages);
+void MeshDeviceImpl::init_realtime_profiler_socket(const std::shared_ptr<MeshDevice>& mesh_device) {
+    if (realtime_profiler_) {
+        return;
     }
-
-    // Enter sync mode - write sync_request = 1
-    std::vector<uint32_t> sync_req_data = {1};
-    tt::tt_metal::detail::WriteToDeviceL1(
-        dev_state.device,
-        dev_state.realtime_profiler_core,
-        dev_state.sync_request_addr,
-        sync_req_data,
-        CoreType::WORKER);
-
-    // Brief pause after entering sync mode to let the kernel process the request
-    std::this_thread::sleep_for(std::chrono::milliseconds(50));
-
-    // Collect samples (+1 for the first one we discard).
-    // Use half-round-trip bracketing: capture host time before the PCIe write and
-    // after the device response, then use the midpoint. This cancels the one-way
-    // PCIe latency bias under the assumption the path is roughly symmetric.
-    constexpr uint32_t kSyncReadTimeoutMs = 2000;
-    uint32_t consecutive_timeouts = 0;
-    constexpr uint32_t kMaxConsecutiveTimeouts = 3;
-
-    for (uint32_t i = 0; i < num_samples + 1; i++) {
-        std::this_thread::sleep_for(std::chrono::milliseconds(5));
-
-        // Send truncated 32-bit value as echo identifier for pairing
-        int64_t host_before = rt_profiler_host_ticks() - host_start_time;
-        uint32_t host_time_id = static_cast<uint32_t>(host_before);
-        std::vector<uint32_t> host_time_data = {host_time_id};
-        tt::tt_metal::detail::WriteToDeviceL1(
-            dev_state.device,
-            dev_state.realtime_profiler_core,
-            dev_state.sync_host_ts_addr,
-            host_time_data,
-            CoreType::WORKER);
-
-        // Poll for response with timeout instead of blocking forever
-        auto deadline = std::chrono::steady_clock::now() + std::chrono::milliseconds(kSyncReadTimeoutMs);
-        bool got_response = false;
-        while (std::chrono::steady_clock::now() < deadline) {
-            if (dev_state.socket->pages_available() > 0) {
-                got_response = true;
-                break;
-            }
-            std::this_thread::sleep_for(std::chrono::microseconds(100));
-        }
-
-        if (!got_response) {
-            consecutive_timeouts++;
-            log_warning(
-                tt::LogMetal,
-                "[Real-time profiler] Device {} sync sample {}/{} timed out after {}ms "
-                "(consecutive timeouts: {}/{})",
-                dev_state.chip_id,
-                i,
-                num_samples,
-                kSyncReadTimeoutMs,
-                consecutive_timeouts,
-                kMaxConsecutiveTimeouts);
-            if (consecutive_timeouts >= kMaxConsecutiveTimeouts) {
-                log_warning(
-                    tt::LogMetal,
-                    "[Real-time profiler] Device {} sync aborted: {} consecutive timeouts. "
-                    "Device kernel may not be responding (check DPRINT output).",
-                    dev_state.chip_id,
-                    consecutive_timeouts);
-                break;
-            }
-            continue;
-        }
-
-        consecutive_timeouts = 0;
-        std::vector<uint32_t> sync_data(kSyncPageWords);
-        dev_state.socket->read(sync_data.data(), 1);
-        // Extract device timestamp and echoed host timestamp
-        uint64_t device_time = (static_cast<uint64_t>(sync_data[0]) << 32) | sync_data[1];
-        uint32_t echoed_host_time = sync_data[2];
-        uint32_t marker = sync_data[3];
-
-        // Discard first sample - can be very off due to cold PCIe path
-        if (i == 0) {
-            continue;
-        }
-
-        // Verify echo matches and marker is valid; use host_before directly
-        // (not midpoint) because the H2D PCIe write (~2µs) and D2H socket
-        // return (~50-80µs) have vastly different latencies, making the
-        // midpoint biased. host_before is ~2µs before the device captures
-        // its wall clock, giving consistent ~2µs accuracy.
-        if (marker == REALTIME_PROFILER_SYNC_MARKER_ID && echoed_host_time == host_time_id) {
-            samples.push_back({host_before, device_time});
-        }
-    }
-
-    // Exit sync mode - write sync_request = 0
-    sync_req_data[0] = 0;
-    tt::tt_metal::detail::WriteToDeviceL1(
-        dev_state.device,
-        dev_state.realtime_profiler_core,
-        dev_state.sync_request_addr,
-        sync_req_data,
-        CoreType::WORKER);
-
-    // Compute sync parameters using numerically stable centered linear regression.
-    //
-    // Model: device_time = slope * host_time + intercept
-    //   where slope = frequency * tracy_ratio  (device_cycles per TSC_tick)
-    //
-    // The standard normal equations suffer from catastrophic cancellation when
-    // computing (n*Σxy - Σx*Σy) with large absolute values (~10^25) to get a
-    // result of ~10^19, losing ~6 digits of precision with double's 15.9 digits.
-    //
-    // Centering: slope = Σ((x - x̄)(y - ȳ)) / Σ((x - x̄)²)
-    // reduces operand magnitudes from ~10^12 to ~10^9 (deviations from mean),
-    // keeping all intermediate products well within double precision.
-    if (samples.size() >= 2) {
-        const double n = static_cast<double>(samples.size());
-        const double tracy_ratio = rt_profiler_ns_per_tick();
-
-        // Pass 1: compute means for centering
-        double host_mean = 0.0, device_mean = 0.0;
-        for (const auto& s : samples) {
-            host_mean += static_cast<double>(s.host_time);
-            device_mean += static_cast<double>(s.device_time);
-        }
-        host_mean /= n;
-        device_mean /= n;
-
-        // Pass 2: centered regression — numerically stable
-        double num = 0.0, den = 0.0;
-        for (const auto& s : samples) {
-            double dx = static_cast<double>(s.host_time) - host_mean;
-            double dy = static_cast<double>(s.device_time) - device_mean;
-            num += dx * dy;
-            den += dx * dx;
-        }
-
-        if (std::abs(den) > 1e-10) {
-            // slope = device_cycles per host_TSC_tick
-            // frequency = slope / tracy_ratio = device_cycles per nanosecond (GHz)
-            double slope = num / den;
-            dev_state.sync_frequency = slope / tracy_ratio;
-        } else {
-            // Fallback to device AICLK
-            dev_state.sync_frequency = cluster.get_device_aiclk(dev_state.chip_id) / 1000.0;
-        }
-
-        // Intercept via means: intercept = ȳ - slope * x̄
-        // This is the device cycle count at host_time = 0 (i.e. at host_start_time)
-        double slope = dev_state.sync_frequency * tracy_ratio;
-        double intercept = device_mean - slope * host_mean;
-        dev_state.first_timestamp = static_cast<uint64_t>(intercept);
-        dev_state.sync_host_start = host_start_time;
-
-        log_info(
-            tt::LogMetal,
-            "[Real-time profiler] Device {} sync complete: {} samples, frequency={:.6f} GHz, "
-            "device_time_at_sync={} cycles",
-            dev_state.chip_id,
-            samples.size(),
-            dev_state.sync_frequency,
-            dev_state.first_timestamp);
-    } else {
-        // Fallback if not enough samples
-        dev_state.sync_frequency = cluster.get_device_aiclk(dev_state.chip_id) / 1000.0;
-        dev_state.first_timestamp = 0;
-        dev_state.sync_host_start = host_start_time;
-        log_warning(
-            tt::LogMetal,
-            "[Real-time profiler] Device {} sync failed - not enough samples, using default frequency",
-            dev_state.chip_id);
-    }
+    realtime_profiler_ = std::make_unique<RealtimeProfilerManager>(mesh_device);
 }
 
 void MeshDeviceImpl::trigger_realtime_profiler_sync_check() {
-    if (realtime_profiler_devices_.empty() || !realtime_profiler_tracy_handler_) {
-        return;
+    if (realtime_profiler_) {
+        realtime_profiler_->trigger_sync_check();
     }
-
-    constexpr uint32_t kPageSize = 64;
-    constexpr uint32_t kPageWords = kPageSize / sizeof(uint32_t);
-    constexpr uint32_t kSyncTimeoutMs = 5000;
-    constexpr uint32_t kPauseTimeoutMs = 2000;
-
-    // 1. Pause the receiver thread so we get exclusive access to the sockets.
-    //    This avoids the GIL deadlock: the receiver thread may be blocked trying
-    //    to acquire the GIL for a Python callback while the calling thread (which
-    //    holds the GIL) waits here. If the receiver doesn't pause in time, we
-    //    cannot safely read from the socket, so we skip the sync check.
-    realtime_profiler_pause_requested_.store(true, std::memory_order_release);
-    {
-        auto pause_deadline = std::chrono::steady_clock::now() + std::chrono::milliseconds(kPauseTimeoutMs);
-        while (!realtime_profiler_paused_.load(std::memory_order_acquire)) {
-            if (std::chrono::steady_clock::now() > pause_deadline) {
-                log_warning(
-                    tt::LogMetal, "[Real-time profiler] Could not pause receiver thread for sync check - skipping");
-                realtime_profiler_pause_requested_.store(false, std::memory_order_release);
-                return;
-            }
-            std::this_thread::sleep_for(std::chrono::microseconds(100));
-        }
-    }
-
-    // Receiver thread is now paused - we have exclusive socket access.
-    std::vector<uint32_t> page_buf(kPageWords);
-
-    for (auto& dev_state : realtime_profiler_devices_) {
-        // 2. Drain any pending data pages so the kernel's socket buffer has room
-        //    for the sync response. Process each page the same way the receiver
-        //    thread would (invoke callbacks, push Tracy markers).
-        while (dev_state.socket->pages_available() > 0) {
-            dev_state.socket->read(page_buf.data(), 1);
-            uint32_t* rp = page_buf.data();
-            uint64_t start_time = (static_cast<uint64_t>(rp[0]) << 32) | rp[1];
-            uint32_t start_id = rp[2];
-            uint64_t end_time = (static_cast<uint64_t>(rp[4]) << 32) | rp[5];
-            if (start_id != 0) {
-                tt::ProgramRealtimeRecord record;
-                record.program_id = start_id;
-                record.start_timestamp = start_time;
-                record.end_timestamp = end_time;
-                record.frequency = dev_state.sync_frequency;
-                record.chip_id = dev_state.chip_id;
-                record.kernel_sources = tt::GetKernelSourcesVecForRuntimeId(start_id);
-                tt::InvokeProgramRealtimeProfilerCallbacks(record);
-            }
-        }
-
-        // 3. Enter sync mode on the device kernel.
-        std::vector<uint32_t> sync_req = {1};
-        tt::tt_metal::detail::WriteToDeviceL1(
-            dev_state.device,
-            dev_state.realtime_profiler_core,
-            dev_state.sync_request_addr,
-            sync_req,
-            CoreType::WORKER);
-
-        std::this_thread::sleep_for(std::chrono::milliseconds(5));
-
-        // 4. Send host timestamp to trigger device response.
-        // Place the host-side Tracy marker right before the PCIe write so both
-        // FINISH_SYNC and SYNC_CHECK use the same timing convention.
-        dev_state.sync_host_time_before = rt_profiler_host_ticks();
-        uint32_t host_time_id = static_cast<uint32_t>(dev_state.sync_host_time_before & 0xFFFFFFFF);
-        std::vector<uint32_t> host_time_data = {host_time_id};
-        TracyMessageL("FINISH_SYNC");
-        tt::tt_metal::detail::WriteToDeviceL1(
-            dev_state.device,
-            dev_state.realtime_profiler_core,
-            dev_state.sync_host_ts_addr,
-            host_time_data,
-            CoreType::WORKER);
-
-        // 5. Read directly from the socket until we get the sync response
-        //    (or timeout). Any data pages that arrive before the sync response
-        //    are processed inline.
-        bool got_sync = false;
-        auto deadline = std::chrono::steady_clock::now() + std::chrono::milliseconds(kSyncTimeoutMs);
-        while (!got_sync) {
-            if (std::chrono::steady_clock::now() > deadline) {
-                log_warning(tt::LogMetal, "[Real-time profiler] Sync check timed out for device {}", dev_state.chip_id);
-                break;
-            }
-
-            if (dev_state.socket->pages_available() == 0) {
-                std::this_thread::sleep_for(std::chrono::microseconds(100));
-                continue;
-            }
-
-            dev_state.socket->read(page_buf.data(), 1);
-            uint32_t* rp = page_buf.data();
-            uint32_t marker = rp[3];
-
-            if (marker == REALTIME_PROFILER_SYNC_MARKER_ID) {
-                uint64_t device_time = (static_cast<uint64_t>(rp[0]) << 32) | rp[1];
-                realtime_profiler_tracy_handler_->CalibrateDevice(
-                    dev_state.chip_id, dev_state.sync_host_time_before, device_time, dev_state.sync_frequency);
-                realtime_profiler_tracy_handler_->PushSyncCheckMarker(
-                    dev_state.chip_id, device_time, dev_state.sync_frequency);
-                got_sync = true;
-            } else {
-                uint64_t start_time = (static_cast<uint64_t>(rp[0]) << 32) | rp[1];
-                uint32_t start_id = rp[2];
-                uint64_t end_time = (static_cast<uint64_t>(rp[4]) << 32) | rp[5];
-                if (start_id != 0) {
-                    tt::ProgramRealtimeRecord record;
-                    record.program_id = start_id;
-                    record.start_timestamp = start_time;
-                    record.end_timestamp = end_time;
-                    record.frequency = dev_state.sync_frequency;
-                    record.chip_id = dev_state.chip_id;
-                    record.kernel_sources = tt::GetKernelSourcesVecForRuntimeId(start_id);
-                    tt::InvokeProgramRealtimeProfilerCallbacks(record);
-                }
-            }
-        }
-
-        // 6. Exit sync mode.
-        sync_req[0] = 0;
-        tt::tt_metal::detail::WriteToDeviceL1(
-            dev_state.device,
-            dev_state.realtime_profiler_core,
-            dev_state.sync_request_addr,
-            sync_req,
-            CoreType::WORKER);
-    }
-
-    // 7. Resume receiver thread.
-    realtime_profiler_pause_requested_.store(false, std::memory_order_release);
-}
-
-void MeshDeviceImpl::init_realtime_profiler_socket(const std::shared_ptr<MeshDevice>& mesh_device) {
-    if (!realtime_profiler_devices_.empty()) {
-        return;
-    }
-
-    // HAL offsets are the same for all devices (same arch)
-    const auto& hal = MetalContext::instance().hal();
-    const auto& factory = hal.get_dev_msgs_factory(HalProgrammableCoreType::TENSIX);
-    // realtime_profiler_msg_t lives in a dispatch-core-local L1 region assigned by
-    // CommandQueueDeviceAddrType::REALTIME_PROFILER_MSG; it is only reachable on the dispatch
-    // cores and the reserved RT-profiler tensix core. Per-field offsets are still computed off
-    // realtime_profiler_msg_t itself.
-    const auto& dispatch_mem_map = MetalContext::instance().dispatch_mem_map();
-    const uint32_t realtime_profiler_base_addr =
-        dispatch_mem_map.get_device_command_queue_addr(CommandQueueDeviceAddrType::REALTIME_PROFILER_MSG);
-    // RealtimeProfilerCoreL1 (ring buffer + D2H socket sender config) lives directly past the
-    // dispatch carve-outs on the reserved profiler tensix. The user-space allocator never lands
-    // here because that core is excluded from the L1 bank table; sizing fits inside one L1 bank
-    // is enforced by evaluate_realtime_profiler_eligibility.
-    const uint32_t rt_profiler_core_l1_base =
-        dispatch_mem_map.get_device_command_queue_addr(CommandQueueDeviceAddrType::UNRESERVED);
-    const auto rt_profiler_core_l1_addrs = compute_rt_profiler_core_l1_addrs(rt_profiler_core_l1_base);
-
-    // The socket sender config region inside RealtimeProfilerCoreL1 must be at least as
-    // large as the D2HSocket implementation actually needs. RT_PROFILER_SOCKET_CONFIG_SIZE
-    // includes headroom (128 B vs. ~64 B today), but if SocketSenderSize ever grows past
-    // it we want a deterministic startup failure rather than a silent L1 overrun.
-    TT_FATAL(
-        RT_PROFILER_SOCKET_CONFIG_SIZE >= D2HSocket::required_config_buffer_size(),
-        "RT_PROFILER_SOCKET_CONFIG_SIZE ({} B) is smaller than D2HSocket's required config "
-        "buffer size ({} B). Bump RT_PROFILER_SOCKET_CONFIG_SIZE in "
-        "tt_metal/impl/dispatch/kernels/realtime_profiler_ring_buffer.hpp and rebuild.",
-        RT_PROFILER_SOCKET_CONFIG_SIZE,
-        D2HSocket::required_config_buffer_size());
-    uint32_t config_buffer_addr_offset = factory.offset_of<dev_msgs::realtime_profiler_msg_t>(
-        dev_msgs::realtime_profiler_msg_t::Field::config_buffer_addr);
-    uint32_t sync_request_offset =
-        factory.offset_of<dev_msgs::realtime_profiler_msg_t>(dev_msgs::realtime_profiler_msg_t::Field::sync_request);
-    uint32_t sync_host_timestamp_offset = factory.offset_of<dev_msgs::realtime_profiler_msg_t>(
-        dev_msgs::realtime_profiler_msg_t::Field::sync_host_timestamp);
-    uint32_t realtime_profiler_mailbox_addr = realtime_profiler_base_addr + config_buffer_addr_offset;
-
-    auto& dispatch_core_manager = MetalContext::instance().get_dispatch_core_manager();
-    const std::string realtime_profiler_kernel_path = "tt_metal/impl/dispatch/kernels/cq_realtime_profiler.cpp";
-    const std::string realtime_profiler_push_kernel_path =
-        "tt_metal/impl/dispatch/kernels/cq_realtime_profiler_push.cpp";
-
-    // Set up real-time profiler for each local device in the mesh
-    for (const auto& coord : MeshCoordinateRange(view_->shape())) {
-        if (!this->is_local(coord)) {
-            continue;
-        }
-
-        IDevice* device = this->get_device(coord);
-        auto device_id = device->id();
-
-        // Single, consolidated eligibility gate for the real-time profiler. Covers
-        // MMIO capability, IOMMU / hugepage fallback preconditions, and the reserved
-        // dispatch-pool tensix core. Logs the specific reason when disabling.
-        auto eligibility = evaluate_realtime_profiler_eligibility(device);
-        if (!eligibility.enabled) {
-            continue;
-        }
-        CoreCoord realtime_profiler_core = eligibility.core;
-
-        log_info(
-            tt::LogMetal,
-            "Using reserved tensix ({}, {}) for real-time profiler on device {}",
-            realtime_profiler_core.x,
-            realtime_profiler_core.y,
-            device_id);
-
-        // Create per-device state
-        RealtimeProfilerDeviceState dev_state;
-        dev_state.device = device;
-        dev_state.chip_id = device_id;
-        dev_state.mesh_coord = coord;
-        dev_state.realtime_profiler_core = realtime_profiler_core;
-        // RT-profiler L1 carve-out (ring + socket config) on this core. All sub-addresses
-        // are derived from a single base anchored at dispatch_mem_map's UNRESERVED, so the
-        // kernels and the D2H socket sender config see a deterministic layout that's not
-        // routed through the user-space allocator.
-        dev_state.core_l1 = rt_profiler_core_l1_addrs;
-
-        // Create D2H socket for this device
-        auto sender_core = MeshCoreCoord{coord, realtime_profiler_core};
-
-        log_info(
-            tt::LogMetal,
-            "Initializing real-time profiler D2H socket for device {} on MeshDevice {}",
-            device_id,
-            this->id());
-
-        // Defensive guard: the D2H socket construction is where host-side memory
-        // pinning / hugepage mapping actually happens, and that interaction with
-        // UMD / the kernel driver has historically been the most fragile part of
-        // bringing up the RT profiler (e.g. `pin_pages` failing on BH without
-        // IOMMU). The `evaluate_realtime_profiler_eligibility` gate above should
-        // already filter out the known-bad configurations, but we don't want a
-        // *future* regression in that path to kill an otherwise-healthy run.
-        // If anything goes wrong here, log a clear warning and fall through to
-        // the next device — no activation is notified for this chip, so the
-        // rest of the stack will simply see "RT profiler not active" for it.
-        try {
-            // Provide the socket sender config L1 address from the dispatch carve-out so
-            // D2HSocket does NOT allocate L1 through MeshBuffer::create on a reserved
-            // dispatch core (which would otherwise be tracked by ttnn graph capture and
-            // crash get_buffer_pages on dispatch cores not in the L1 bank table).
-            dev_state.socket = std::make_unique<D2HSocket>(
-                mesh_device,
-                sender_core,
-                RealtimeProfilerRuntimeSizes::fifo_size,
-                D2HSocket::ExternalConfigBuffer{.address = dev_state.core_l1.socket_config});
-            dev_state.socket->set_page_size(RealtimeProfilerRuntimeSizes::page_size);
-        } catch (const std::exception& e) {
-            log_warning(
-                tt::LogMetal,
-                "Real-time profiler disabled on device {}: D2H socket construction failed ({}). "
-                "This typically indicates a host-side memory pinning / hugepage mapping issue "
-                "(e.g. IOMMU misconfiguration or UMD DMA pin failure). Continuing without RT "
-                "profiler on this device.",
-                device_id,
-                e.what());
-            continue;
-        }
-
-        // Store sync and diagnostic addresses
-        dev_state.realtime_profiler_base_addr = realtime_profiler_base_addr;
-        dev_state.sync_request_addr = realtime_profiler_base_addr + sync_request_offset;
-        dev_state.sync_host_ts_addr = realtime_profiler_base_addr + sync_host_timestamp_offset;
-
-        // Write real-time profiler core info to dispatch_s core's mailbox for termination signaling
-        if (dispatch_core_manager.is_dispatcher_s_core_allocated(device_id, 0, 0)) {
-            const tt_cxy_pair& dispatch_s_cxy = dispatch_core_manager.dispatcher_s_core(device_id, 0, 0);
-            CoreCoord dispatch_s_core(dispatch_s_cxy.x, dispatch_s_cxy.y);
-
-            CoreCoord realtime_profiler_virtual =
-                device->virtual_core_from_logical_core(realtime_profiler_core, CoreType::WORKER);
-            uint32_t realtime_profiler_noc_xy =
-                hal.noc_xy_encoding(realtime_profiler_virtual.x, realtime_profiler_virtual.y);
-
-            uint32_t realtime_profiler_core_noc_xy_offset = factory.offset_of<dev_msgs::realtime_profiler_msg_t>(
-                dev_msgs::realtime_profiler_msg_t::Field::realtime_profiler_core_noc_xy);
-            uint32_t realtime_profiler_mailbox_addr_offset = factory.offset_of<dev_msgs::realtime_profiler_msg_t>(
-                dev_msgs::realtime_profiler_msg_t::Field::realtime_profiler_mailbox_addr);
-            uint32_t realtime_profiler_state_offset = factory.offset_of<dev_msgs::realtime_profiler_msg_t>(
-                dev_msgs::realtime_profiler_msg_t::Field::realtime_profiler_state);
-            uint32_t realtime_profiler_core_state_addr = realtime_profiler_base_addr + realtime_profiler_state_offset;
-            uint32_t dispatch_s_mailbox_base = realtime_profiler_base_addr;
-
-            std::vector<uint32_t> noc_xy_data = {realtime_profiler_noc_xy};
-            tt::tt_metal::detail::WriteToDeviceL1(
-                device,
-                dispatch_s_core,
-                dispatch_s_mailbox_base + realtime_profiler_core_noc_xy_offset,
-                noc_xy_data,
-                CoreType::WORKER);
-
-            std::vector<uint32_t> mailbox_addr_data = {realtime_profiler_core_state_addr};
-            tt::tt_metal::detail::WriteToDeviceL1(
-                device,
-                dispatch_s_core,
-                dispatch_s_mailbox_base + realtime_profiler_mailbox_addr_offset,
-                mailbox_addr_data,
-                CoreType::WORKER);
-
-            log_info(
-                tt::LogMetal,
-                "Device {}: wrote real-time profiler core info (noc_xy=0x{:x}, mailbox_addr=0x{:x}) to dispatch_s ({}, "
-                "{})",
-                device_id,
-                realtime_profiler_noc_xy,
-                realtime_profiler_core_state_addr,
-                dispatch_s_core.x,
-                dispatch_s_core.y);
-        }
-
-        // Ring buffer (BRISC->NCRISC handoff) lives at a fixed offset inside the
-        // RT-profiler L1 carve-out. It is *not* created via Buffer::create / MeshBuffer
-        // because the profiler core is reserved out of the user-space allocator's L1 bank
-        // table; allocating here would either fail with "No L1 bank exists" or — worse —
-        // get tracked by ttnn graph capture and crash get_buffer_pages later.
-        const uint32_t ring_buffer_addr = dev_state.core_l1.ring_buffer;
-
-        // Get PCIe core NOC-0 coordinates for WH (NCRISC kernel translates to NOC 1)
-        uint32_t pcie_noc_x = 0;
-        uint32_t pcie_noc_y = 0;
-        bool need_pcie_noc_defines = false;
-        {
-            const auto& cluster = MetalContext::instance().get_cluster();
-            auto arch = MetalContext::instance().hal().get_arch();
-            if (arch == tt::ARCH::WORMHOLE_B0) {
-                ChipId mmio_device_id = cluster.get_associated_mmio_device(device_id);
-                const auto& soc = cluster.get_soc_desc(mmio_device_id);
-                const auto& pcie_cores = soc.get_cores(CoreType::PCIE, CoordSystem::NOC0);
-                TT_ASSERT(!pcie_cores.empty());
-                pcie_noc_x = pcie_cores.front().x;
-                pcie_noc_y = pcie_cores.front().y;
-                need_pcie_noc_defines = true;
-            }
-        }
-        // Zero the ring buffer header (write_index, read_index, terminate, _pad[13])
-        // to clear stale state from a previous run (e.g. terminate=1 from teardown).
-        // The header is everything before RtProfilerRingBuffer::data, so use offsetof
-        // to track the header size automatically if the struct grows.
-        {
-            constexpr uint32_t kRingHeaderBytes = offsetof(RtProfilerRingBuffer, data);
-            static_assert(kRingHeaderBytes % sizeof(uint32_t) == 0, "Ring header must be uint32-aligned");
-            std::vector<uint32_t> zero_header(kRingHeaderBytes / sizeof(uint32_t), 0);
-            tt::tt_metal::detail::WriteToDeviceL1(
-                device, realtime_profiler_core, ring_buffer_addr, zero_header, CoreType::WORKER);
-        }
-
-        // Zero the entire realtime_profiler_msg_t region in L1 BEFORE launching the
-        // kernels. L1 is not guaranteed to be zero-initialized between runs, so the
-        // profiler mailbox can carry stale values across a MeshDevice re-create.
-        // Any of these fields being non-zero at BRISC/NCRISC boot can misbehave:
-        //   * config_buffer_addr != 0  -> NCRISC exits its config wait loop
-        //     immediately and reads garbage socket config.
-        //   * sync_request != 0        -> BRISC enters realtime_profiler_sync() at
-        //     boot before the host has reached the sync phase.
-        //   * sync_host_timestamp != 0 -> BRISC pushes a stale sync marker to the
-        //     ring buffer on first boot; NCRISC forwards it as a phantom page.
-        //   * realtime_profiler_state / program_id_fifo_{start,end} stale values
-        //     similarly corrupt the kernel state machine on first boot.
-        // realtime_profiler_base_addr == realtime_profiler_mailbox_addr (the mailbox
-        // base is the start of the struct), so zero from realtime_profiler_base_addr.
-        {
-            const uint32_t profiler_msg_size = factory.size_of<dev_msgs::realtime_profiler_msg_t>();
-            const uint32_t profiler_msg_words = profiler_msg_size / sizeof(uint32_t);
-            std::vector<uint32_t> zero_msg(profiler_msg_words, 0);
-            tt::tt_metal::detail::WriteToDeviceL1(
-                device, realtime_profiler_core, realtime_profiler_base_addr, zero_msg, CoreType::WORKER);
-        }
-
-        // Compile and launch real-time profiler kernels (BRISC reader + NCRISC pusher)
-        {
-            Program realtime_profiler_program;
-
-            uint32_t dispatch_core_noc_x = 0;
-            uint32_t dispatch_core_noc_y = 0;
-            uint32_t dispatch_data_addr_a = 0;
-            uint32_t dispatch_data_addr_b = 0;
-            if (dispatch_core_manager.is_dispatcher_s_core_allocated(device_id, 0, 0)) {
-                const tt_cxy_pair& dispatch_s_cxy = dispatch_core_manager.dispatcher_s_core(device_id, 0, 0);
-                CoreCoord dispatch_s_virtual = device->virtual_core_from_logical_core(
-                    CoreCoord(dispatch_s_cxy.x, dispatch_s_cxy.y), CoreType::WORKER);
-                dispatch_core_noc_x = dispatch_s_virtual.x;
-                dispatch_core_noc_y = dispatch_s_virtual.y;
-
-                uint32_t kernel_start_a_offset = factory.offset_of<dev_msgs::realtime_profiler_msg_t>(
-                    dev_msgs::realtime_profiler_msg_t::Field::kernel_start_a);
-                uint32_t kernel_start_b_offset = factory.offset_of<dev_msgs::realtime_profiler_msg_t>(
-                    dev_msgs::realtime_profiler_msg_t::Field::kernel_start_b);
-                dispatch_data_addr_a = realtime_profiler_base_addr + kernel_start_a_offset;
-                dispatch_data_addr_b = realtime_profiler_base_addr + kernel_start_b_offset;
-            }
-
-            // BRISC kernel: reads from dispatch_s, writes to ring buffer
-            DataMovementConfig brisc_config;
-            brisc_config.processor = DataMovementProcessor::RISCV_0;
-            brisc_config.noc = NOC::RISCV_0_default;
-            brisc_config.defines["DISPATCH_CORE_NOC_X"] = std::to_string(dispatch_core_noc_x);
-            brisc_config.defines["DISPATCH_CORE_NOC_Y"] = std::to_string(dispatch_core_noc_y);
-            brisc_config.defines["DISPATCH_DATA_ADDR_A"] = std::to_string(dispatch_data_addr_a);
-            brisc_config.defines["DISPATCH_DATA_ADDR_B"] = std::to_string(dispatch_data_addr_b);
-            brisc_config.defines["RING_BUFFER_ADDR"] = std::to_string(ring_buffer_addr);
-            brisc_config.defines["REALTIME_PROFILER_MSG_ADDR"] = std::to_string(realtime_profiler_base_addr);
-            CreateKernel(
-                realtime_profiler_program, realtime_profiler_kernel_path, realtime_profiler_core, brisc_config);
-
-            // NCRISC kernel: drains ring buffer to host via PCIe
-            DataMovementConfig ncrisc_config;
-            ncrisc_config.processor = DataMovementProcessor::RISCV_1;
-            ncrisc_config.noc = NOC::RISCV_1_default;
-            ncrisc_config.defines["RING_BUFFER_ADDR"] = std::to_string(ring_buffer_addr);
-            ncrisc_config.defines["REALTIME_PROFILER_MSG_ADDR"] = std::to_string(realtime_profiler_base_addr);
-            if (need_pcie_noc_defines) {
-                ncrisc_config.defines["RT_PROFILER_PCIE_NOC_X"] = std::to_string(pcie_noc_x);
-                ncrisc_config.defines["RT_PROFILER_PCIE_NOC_Y"] = std::to_string(pcie_noc_y);
-            }
-            CreateKernel(
-                realtime_profiler_program, realtime_profiler_push_kernel_path, realtime_profiler_core, ncrisc_config);
-
-            tt::tt_metal::detail::CompileProgram(device, realtime_profiler_program, /*force_slow_dispatch=*/true);
-            ::tt::tt_metal::detail::WriteRuntimeArgsToDevice(
-                device, realtime_profiler_program, /*force_slow_dispatch=*/true);
-            ::tt::tt_metal::detail::LaunchProgram(
-                device, realtime_profiler_program, /*wait_until_cores_done=*/false, /*force_slow_dispatch=*/true);
-
-            // realtime_profiler_msg_t lives outside mailboxes_t (see
-            // CommandQueueDeviceAddrType::REALTIME_PROFILER_MSG), so LaunchProgram's writes to
-            // launch messages / go signals do not overlap with config_buffer_addr. The write is
-            // performed here to keep ordering deterministic relative to launch.
-            uint32_t config_buffer_addr = dev_state.socket->get_config_buffer_address();
-            std::vector<uint32_t> addr_data = {config_buffer_addr};
-            tt::tt_metal::detail::WriteToDeviceL1(
-                device, realtime_profiler_core, realtime_profiler_mailbox_addr, addr_data, CoreType::WORKER);
-
-            log_info(
-                tt::LogMetal,
-                "Device {}: launched real-time profiler BRISC+NCRISC kernels on core ({}, {}), "
-                "ring_buffer_addr=0x{:x}, config_buffer_addr=0x{:x}",
-                device_id,
-                realtime_profiler_core.x,
-                realtime_profiler_core.y,
-                ring_buffer_addr,
-                config_buffer_addr);
-        }
-
-        realtime_profiler_devices_.push_back(std::move(dev_state));
-    }
-
-    if (realtime_profiler_devices_.empty()) {
-        log_warning(
-            tt::LogMetal, "[Real-time profiler] No local devices found in mesh, skipping real-time profiler setup");
-        return;
-    }
-
-    // Announce activation so tt::IsProgramRealtimeProfilerActive() (and its Python
-    // equivalent) can report the true state to external callers. This is the first
-    // point at which RT profiler has live per-chip state on MeshDevice side; we mark
-    // deactivation on close, paired with realtime_profiler_devices_.clear().
-    for (const auto& dev_state : realtime_profiler_devices_) {
-        tt::NotifyProgramRealtimeProfilerActivated(dev_state.chip_id);
-    }
-
-    // Always run the real-time profiler's own host-device sync.
-    // The device profiler's SyncInfo (from TT_METAL_PROFILER_SYNC) cannot be reused here
-    // because syncDeviceHost reconstructs device timestamps with a 12-bit masked high word
-    // (sync_times[0] & 0xFFF), truncating the 64-bit wall clock. The real-time profiler
-    // reads full 64-bit timestamps, so using the truncated anchor would shift all zones
-    // by hours in Tracy.
-    for (auto& dev_state : realtime_profiler_devices_) {
-        constexpr uint32_t kMaxSyncRetries = 3;
-        constexpr uint32_t kRetryDelayMs = 500;
-        for (uint32_t attempt = 0; attempt <= kMaxSyncRetries; attempt++) {
-            if (attempt > 0) {
-                log_info(
-                    tt::LogMetal,
-                    "[Real-time profiler] Device {} sync retry {}/{}",
-                    dev_state.chip_id,
-                    attempt,
-                    kMaxSyncRetries);
-                std::this_thread::sleep_for(std::chrono::milliseconds(kRetryDelayMs));
-            }
-            run_realtime_profiler_sync(dev_state, 100);
-            if (dev_state.first_timestamp != 0) {
-                break;
-            }
-        }
-    }
-
-    // Create Tracy handler and register all device contexts
-    realtime_profiler_tracy_handler_ = std::make_unique<RealtimeProfilerTracyHandler>();
-    for (const auto& dev_state : realtime_profiler_devices_) {
-        realtime_profiler_tracy_handler_->AddDevice(
-            dev_state.chip_id,
-            dev_state.sync_host_start,
-            static_cast<double>(dev_state.first_timestamp),
-            dev_state.sync_frequency);
-    }
-
-    // Emit sync verification markers: take one independent device measurement per device
-    // and push paired host + device events. In Tracy, the horizontal distance between the
-    // host "SYNC_CHECK" zone and the device "SYNC_CHECK" zone is the sync error.
-    for (auto& dev_state : realtime_profiler_devices_) {
-        // Re-enter sync mode
-        std::vector<uint32_t> sync_req = {1};
-        tt::tt_metal::detail::WriteToDeviceL1(
-            dev_state.device,
-            dev_state.realtime_profiler_core,
-            dev_state.sync_request_addr,
-            sync_req,
-            CoreType::WORKER);
-
-        std::this_thread::sleep_for(std::chrono::milliseconds(10));
-
-        // Send a host timestamp to trigger device response.
-        // Same anchor convention as trigger_realtime_profiler_sync_check: capture
-        // host TSC, emit Tracy message, then PCIe write; after the device response
-        // call CalibrateDevice before PushSyncCheckMarker. Without CalibrateDevice,
-        // the device zone is placed using only the coarse regression from
-        // run_realtime_profiler_sync while the host message uses the message
-        // timestamp — extrapolation skew can exceed the sync-accuracy test's ±10µs.
-        int64_t sync_check_host_anchor = rt_profiler_host_ticks();
-        uint32_t host_time_id = 0x5C5C5C5C;  // recognizable pattern
-        std::vector<uint32_t> host_time_data = {host_time_id};
-        TracyMessageL("SYNC_CHECK");
-        tt::tt_metal::detail::WriteToDeviceL1(
-            dev_state.device,
-            dev_state.realtime_profiler_core,
-            dev_state.sync_host_ts_addr,
-            host_time_data,
-            CoreType::WORKER);
-
-        // Wait for device response with timeout
-        constexpr uint32_t kSyncCheckTimeoutMs = 3000;
-        auto sc_deadline = std::chrono::steady_clock::now() + std::chrono::milliseconds(kSyncCheckTimeoutMs);
-        bool sc_got_response = false;
-        while (std::chrono::steady_clock::now() < sc_deadline) {
-            if (dev_state.socket->pages_available() > 0) {
-                sc_got_response = true;
-                break;
-            }
-            std::this_thread::sleep_for(std::chrono::microseconds(100));
-        }
-
-        // Exit sync mode regardless of whether we got a response
-        sync_req[0] = 0;
-        tt::tt_metal::detail::WriteToDeviceL1(
-            dev_state.device,
-            dev_state.realtime_profiler_core,
-            dev_state.sync_request_addr,
-            sync_req,
-            CoreType::WORKER);
-
-        if (sc_got_response) {
-            std::vector<uint32_t> sync_page(RealtimeProfilerRuntimeSizes::page_size / sizeof(uint32_t));
-            dev_state.socket->read(sync_page.data(), 1);
-            uint64_t device_time = (static_cast<uint64_t>(sync_page[0]) << 32) | sync_page[1];
-
-            realtime_profiler_tracy_handler_->CalibrateDevice(
-                dev_state.chip_id, sync_check_host_anchor, device_time, dev_state.sync_frequency);
-            realtime_profiler_tracy_handler_->PushSyncCheckMarker(
-                dev_state.chip_id, device_time, dev_state.sync_frequency);
-
-            log_info(
-                tt::LogMetal,
-                "[Real-time profiler] Device {} sync check: device_time={} cycles",
-                dev_state.chip_id,
-                device_time);
-        } else {
-            log_warning(
-                tt::LogMetal,
-                "[Real-time profiler] Device {} sync check timed out after {}ms, skipping",
-                dev_state.chip_id,
-                kSyncCheckTimeoutMs);
-        }
-    }
-
-    // Start background receiver thread that polls all device sockets round-robin
-    realtime_profiler_stop_.store(false);
-    realtime_profiler_thread_ = std::thread([this]() {
-        tracy::SetThreadName("RealtimeProfiler");
-        uint64_t pages_received = 0;
-
-        log_info(
-            tt::LogMetal,
-            "[Real-time profiler] Receiver thread started for {} devices",
-            realtime_profiler_devices_.size());
-
-        // Helper lambda: process one page from a device socket.
-        // Returns true if a page was consumed, false if nothing was available.
-        std::vector<uint32_t> page_buf(RealtimeProfilerRuntimeSizes::page_size / sizeof(uint32_t));
-        auto process_one_page = [&](RealtimeProfilerDeviceState& dev_state) -> bool {
-            uint32_t available = dev_state.socket->pages_available();
-            if (available == 0) {
-                return false;
-            }
-
-            ZoneScopedN("ProcessPage");
-            dev_state.socket->read(page_buf.data(), 1);
-            uint32_t* read_ptr = page_buf.data();
-
-            // Check for sync response packets only when a sync is pending
-            uint32_t marker = read_ptr[3];
-            if (!dev_state.sync_response_received.load() && marker == REALTIME_PROFILER_SYNC_MARKER_ID) {
-                uint64_t device_time = (static_cast<uint64_t>(read_ptr[0]) << 32) | read_ptr[1];
-                realtime_profiler_tracy_handler_->CalibrateDevice(
-                    dev_state.chip_id, dev_state.sync_host_time_before, device_time, dev_state.sync_frequency);
-                realtime_profiler_tracy_handler_->PushSyncCheckMarker(
-                    dev_state.chip_id, device_time, dev_state.sync_frequency);
-                pages_received++;
-                dev_state.sync_response_received.store(true);
-                return true;
-            }
-
-            // Extract timestamps: kernel_start (words 0-3), kernel_end (words 4-7)
-            // Each realtime_profiler_timestamp_t: time_hi, time_lo, id, header
-            uint64_t start_time = (static_cast<uint64_t>(read_ptr[0]) << 32) | read_ptr[1];
-            uint32_t start_id = read_ptr[2];
-            uint64_t end_time = (static_cast<uint64_t>(read_ptr[4]) << 32) | read_ptr[5];
-
-            // Invoke registered real-time profiler callbacks.
-            // Skip records with id==0: these are non-GO dispatch commands
-            // (e.g. SET_NUM_WORKER_SEMS, SET_GO_SIGNAL_NOC_DATA) that have
-            // no valid program and may contain stale end timestamps.
-            if (start_id != 0) {
-                ZoneScopedN("InvokeCallbacks");
-                tt::ProgramRealtimeRecord record;
-                record.program_id = start_id;
-                record.start_timestamp = start_time;
-                record.end_timestamp = end_time;
-                record.frequency = dev_state.sync_frequency;
-                record.chip_id = dev_state.chip_id;
-                record.kernel_sources = tt::GetKernelSourcesVecForRuntimeId(start_id);
-                tt::InvokeProgramRealtimeProfilerCallbacks(record);
-            }
-
-            pages_received++;
-            return true;
-        };
-
-        while (!realtime_profiler_stop_.load()) {
-            if (realtime_profiler_pause_requested_.load(std::memory_order_acquire)) {
-                realtime_profiler_paused_.store(true, std::memory_order_release);
-                while (realtime_profiler_pause_requested_.load(std::memory_order_acquire) &&
-                       !realtime_profiler_stop_.load()) {
-                    std::this_thread::sleep_for(std::chrono::microseconds(100));
-                }
-                realtime_profiler_paused_.store(false, std::memory_order_release);
-                continue;
-            }
-
-            ZoneScopedN("PollLoop");
-            bool any_data = false;
-
-            for (auto& dev_state : realtime_profiler_devices_) {
-                try {
-                    if (process_one_page(dev_state)) {
-                        any_data = true;
-                    }
-                } catch (const std::exception& e) {
-                    log_warning(
-                        tt::LogMetal,
-                        "[Real-time profiler] Exception in receiver for device {}: {}",
-                        dev_state.chip_id,
-                        e.what());
-                }
-            }
-
-            if (!any_data) {
-                ZoneScopedN("Idle");
-                std::this_thread::sleep_for(std::chrono::milliseconds(10));
-            }
-        }
-
-        // Drain remaining data from sockets. The device may have pushed pages
-        // that arrived after the last poll iteration or while the thread was
-        // sleeping. Keep draining until all sockets are empty for several
-        // consecutive attempts to account for in-flight PCIe writes.
-        {
-            ZoneScopedN("DrainShutdown");
-            constexpr uint32_t kDrainQuietRounds = 10;
-            uint64_t drain_pages = 0;
-            uint32_t quiet_rounds = 0;
-            while (quiet_rounds < kDrainQuietRounds) {
-                bool any_data = false;
-                for (auto& dev_state : realtime_profiler_devices_) {
-                    try {
-                        if (process_one_page(dev_state)) {
-                            any_data = true;
-                            drain_pages++;
-                        }
-                    } catch (const std::exception& e) {
-                        log_warning(
-                            tt::LogMetal,
-                            "[Real-time profiler] Exception draining device {}: {}",
-                            dev_state.chip_id,
-                            e.what());
-                    }
-                }
-                if (any_data) {
-                    quiet_rounds = 0;
-                } else {
-                    quiet_rounds++;
-                    std::this_thread::sleep_for(std::chrono::milliseconds(1));
-                }
-            }
-
-            log_info(
-                tt::LogMetal,
-                "[Real-time profiler] Receiver thread stopped after {} pages ({} drained during shutdown)",
-                pages_received,
-                drain_pages);
-        }
-    });
 }
 
 D2HSocket* MeshDeviceImpl::get_realtime_profiler_socket() const {
-    return realtime_profiler_devices_.empty() ? nullptr : realtime_profiler_devices_.front().socket.get();
+    return realtime_profiler_ ? realtime_profiler_->get_socket() : nullptr;
 }
 
 program_cache::detail::ProgramCache& MeshDeviceImpl::get_program_cache() { return *program_cache_; }
